Move material application to demo page
